Updated package.json to use module rather than main
This is due to the fact that dynamic imports aren't properly
handled by subsequent webpack builds, or at least, I don't know
how to write a webpack config that can deal with this.
There seem to be a few webpack issues related to this such as:
    - webpack/webpack#2471
    - webpack/webpack#6818
    - webpack/webpack#7843

A by-product of this change is that all end-users of the library have
to run webpack or some other bundler that can resolve and organize
the dynamic imports.

Updated package.json to include compiled-js instead of dist
in the list of files to be packed for npm